Refactor to make separate components for instructions and settings.
